Move command-line parsing to a separate file
